Generalize Cache Tests
